refactor(core): lift manifest_helpers from unifi-mcp-shared to unifi-core
Re-export shim left in unifi-mcp-shared for backwards compatibility during Phase 0;
shim removed in Task 17.
